CTCTRADERS-2574 Change view spec to make it easier to test single view
